Create tibble with filenames and timestamps
Information about log files are kept in named lists that can
be added to as more types of log files are brought into the system.
Code creates a tibble containing each filename, the log type, and
the log date that is parsed from the filename using information
contained in the named lists.